+++ app/controllers/reservas_controller.rb
@@ -2,18 +2,21 @@ class ReservasController < InheritedResources::Base
   load_and_authorize_resource
   respond_to :html, :xml, :js, :json
   def index
-    @search = Reserva.search((params[:search] if params[:search]))
-
+    @search = Reserva.with_includes.search(params[:search])
+    @reservas = @search.paginate :page => params[:page], :per_page =>10
     respond_to do |format|
       format.js
-          paginate
       format.html
-          paginate
       format.pdf do
         @reservas = @search
         output = ReservaReport.new.situacion_operadora(@reservas, params[:search])
         send_data output, :filename => "index_report.pdf",
                           :type => "application/pdf"
+      end
+      format.csv do
+        csv = to_csv(@search)
+        send_data csv, :filename => "reservas_#{Time.now}.csv",
+                  :type => "text/csv; charset=utf-8; header=present"
       end
     end
   end
@@ -66,8 +69,17 @@ class ReservasController < InheritedResources::Base
     end
   end
 
-  def paginate
-    @reservas = @search.paginate :page => params[:page], :per_page =>10
+  private
+
+  def to_csv(relation)
+    require 'csv'
+
+    CSV.generate(:col_sep => ";") do |csv|
+      csv << %w[ id agencia operadora fecha salida referencia programa periodo regimen titular pasajeros bruto seguro neto moneda ]
+      relation.each do |r|
+        csv << %W[ #{r.id} #{r.agencia.name} #{r.operadora.name} #{r.fecha} #{r.salida} #{r.referencia} #{r.programa.try(:name)} #{r.periodo} #{r.regimen} #{r.titular} #{r.pasajeros.size} #{r.tarifa.try(:localize)} #{r.seguro.try(:localize)} #{r.total} #{r.total_currency}]
+      end
+    end
   end
 end