+++ db/migrate/20111226224424_add_monto_final_to_deposito.rb
@@ -1,0 +1,12 @@
+class AddMontoFinalToDeposito < ActiveRecord::Migration
+  def self.up
+    add_column :movimientos, :monto_final_cents, :integer
+    add_column :movimientos, :monto_final_currency, :string, :limit => 3
+  end
+
+  def self.down
+    remove_column :movimientos, :monto_final_currency
+    remove_column :movimientos, :monto_final_cents
+  end
+end
+