+++ db/schema.rb
@@ -11,7 +11,7 @@
 #
 # It's strongly recommended to check this file into your version control system.
 
-ActiveRecord::Schema.define(:version => 20110729124450) do
+ActiveRecord::Schema.define(:version => 20120112122545) do
 
   create_table "blogs", :force => true do |t|
     t.string   "title"
@@ -59,7 +59,6 @@ ActiveRecord::Schema.define(:version => 20110729124450) do
   create_table "entidad_versions", :force => true do |t|
     t.integer  "entidad_id"
     t.integer  "version"
-    t.string   "name"
     t.string   "calle"
     t.string   "cuit"
     t.string   "telefono"
@@ -67,6 +66,7 @@ ActiveRecord::Schema.define(:version => 20110729124450) do
     t.string   "email"
     t.string   "web"
     t.integer  "localidad_id"
+    t.string   "name"
     t.datetime "created_at"
     t.datetime "updated_at"
     t.boolean  "hidden",         :default => false
@@ -77,8 +77,6 @@ ActiveRecord::Schema.define(:version => 20110729124450) do
   add_index "entidad_versions", ["entidad_id"], :name => "index_entidad_versions_on_entidad_id"
 
   create_table "entidads", :force => true do |t|
-    t.string   "name"
-    t.string   "type"
     t.string   "calle"
     t.string   "cuit"
     t.string   "telefono"
@@ -86,6 +84,8 @@ ActiveRecord::Schema.define(:version => 20110729124450) do
     t.string   "email"
     t.string   "web"
     t.integer  "localidad_id"
+    t.string   "name"
+    t.string   "type"
     t.datetime "created_at"
     t.datetime "updated_at"
     t.boolean  "hidden",       :default => false
@@ -130,6 +130,25 @@ ActiveRecord::Schema.define(:version => 20110729124450) do
     t.datetime "updated_at"
   end
 
+  create_table "movimiento_versions", :force => true do |t|
+    t.integer  "movimiento_id"
+    t.integer  "version"
+    t.integer  "entidad_id"
+    t.integer  "reserva_id"
+    t.integer  "tdeposito_id"
+    t.integer  "saldo_id"
+    t.integer  "monto_id"
+    t.integer  "numero"
+    t.date     "fecha"
+    t.datetime "created_at"
+    t.datetime "updated_at"
+    t.boolean  "hidden",         :default => false
+    t.integer  "user_id"
+    t.string   "versioned_type"
+  end
+
+  add_index "movimiento_versions", ["movimiento_id"], :name => "index_movimiento_versions_on_movimiento_id"
+
   create_table "movimientos", :force => true do |t|
     t.integer  "entidad_id"
     t.integer  "reserva_id"
@@ -141,14 +160,21 @@ ActiveRecord::Schema.define(:version => 20110729124450) do
     t.date     "fecha"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.boolean  "hidden",                      :default => false
+    t.boolean  "hidden",                            :default => false
     t.integer  "user_id"
+    t.integer  "version"
     t.integer  "operadora_id"
     t.integer  "monto_cents"
-    t.string   "monto_currency", :limit => 3
+    t.string   "monto_currency",       :limit => 3
     t.integer  "movimiento_id"
     t.string   "observaciones"
+    t.integer  "monto_final_cents"
+    t.string   "monto_final_currency", :limit => 3
   end
+
+  add_index "movimientos", ["entidad_id"], :name => "index_movimientos_on_entidad_id"
+  add_index "movimientos", ["reserva_id"], :name => "index_movimientos_on_reserva_id"
+  add_index "movimientos", ["type"], :name => "index_movimientos_on_type"
 
   create_table "pages", :force => true do |t|
     t.string   "title"
@@ -251,8 +277,8 @@ ActiveRecord::Schema.define(:version => 20110729124450) do
     t.float    "seguro"
     t.integer  "monto_id"
     t.string   "referencia"
-    t.boolean  "cancelada",                      :default => false
-    t.boolean  "activa",                         :default => false
+    t.boolean  "cancelada",                         :default => false
+    t.boolean  "activa",                            :default => false
     t.float    "tarifa"
     t.float    "float"
     t.float    "aereo"
@@ -260,11 +286,12 @@ ActiveRecord::Schema.define(:version => 20110729124450) do
     t.float    "pago_minimo"
     t.datetime "created_at"
     t.datetime "updated_at"
-    t.boolean  "hidden",                         :default => false
+    t.boolean  "hidden",                            :default => false
     t.integer  "user_id"
+    t.string   "reserva_versionscol", :limit => 45
     t.date     "voucher"
     t.integer  "total_cents"
-    t.string   "total_currency",    :limit => 3
+    t.string   "total_currency",      :limit => 3
     t.integer  "iva_cents"
     t.integer  "tarifa_cents"
     t.integer  "aereo_cents"
@@ -272,8 +299,8 @@ ActiveRecord::Schema.define(:version => 20110729124450) do
     t.integer  "pago_minimo_cents"
     t.integer  "impuesto_cents"
     t.integer  "seguro_cents"
-    t.boolean  "liquido_agencia",                :default => false
-    t.boolean  "liquido_operadora",              :default => false
+    t.boolean  "liquido_agencia",                   :default => false
+    t.boolean  "liquido_operadora",                 :default => false
     t.integer  "comision"
     t.integer  "habitaciones"
   end
@@ -325,6 +352,10 @@ ActiveRecord::Schema.define(:version => 20110729124450) do
     t.integer  "comision"
     t.integer  "habitaciones"
   end
+
+  add_index "reservas", ["agency_id"], :name => "index_reservas_on_agency_id"
+  add_index "reservas", ["operadora_id"], :name => "index_reservas_on_operadora_id"
+  add_index "reservas", ["programa_id"], :name => "index_reservas_on_programa_id"
 
   create_table "roles", :force => true do |t|
     t.string   "name"
@@ -437,10 +468,8 @@ ActiveRecord::Schema.define(:version => 20110729124450) do
   add_index "users", ["username"], :name => "index_users_on_username", :unique => true
 
   create_table "viajeros", :id => false, :force => true do |t|
-    t.integer  "reserva_id"
-    t.integer  "pasajero_id"
-    t.datetime "created_at"
-    t.datetime "updated_at"
+    t.integer "reserva_id"
+    t.integer "pasajero_id"
   end
 
 end